# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Sociality API",
    "version": "1.0.0",
    "description": "REST API Sociality - Clone from original API"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "tags": [
    { "name": "Auth" },
    { "name": "My Profile" },
    { "name": "Users" },
    { "name": "Posts" },
    { "name": "Likes" },
    { "name": "Comments" },
    { "name": "Follow" },
    { "name": "Saves" }
  ],
  "paths": {
    "/api/auth/register": {
      "post": {
        "summary": "Register new user",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "username", "email", "password"],
                "properties": {
                  "name": { "type": "string", "example": "John Doe" },
                  "username": { "type": "string", "example": "johndoe" },
                  "email": { "type": "string", "example": "john@email.com" },
                  "phone": { "type": "string", "example": "081234567890" },
                  "password": {
                    "type": "string",
                    "minLength": 6,
                    "example": "password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created" },
          "400": { "description": "Bad Request" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Login user",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "example": "john@email.com" },
                  "password": { "type": "string", "example": "password123" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "OK" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/me": {
      "get": {
        "summary": "Get my profile + quick stats",
        "tags": ["My Profile"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "OK" } }
      },
      "patch": {
        "summary": "Update my basic profile (supports avatar upload)",
        "tags": ["My Profile"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "bio": { "type": "string" },
                  "phone": { "type": "string" },
                  "image": { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "OK" },
          "400": { "description": "Bad Request" }
        }
      }
    },
    "/api/me/posts": {
      "get": {
        "summary": "Get posts created by the authenticated user",
        "tags": ["Posts"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": { "type": "integer", "default": 20, "maximum": 50 }
          }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/feed": {
      "get": {
        "summary": "Get feed (posts from followed users)",
        "tags": ["Posts"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": { "type": "integer", "default": 20, "maximum": 50 }
          }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/users/search": {
      "get": {
        "tags": ["Users"],
        "summary": "Search users by name/username",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": { "type": "string", "minLength": 1 },
            "description": "Keyword untuk name atau username"
          },
          {
            "in": "query",
            "name": "page",
            "schema": { "type": "integer", "default": 1, "minimum": 1 }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": { "type": "integer", "default": 20, "maximum": 50 }
          }
        ],
        "responses": { "200": { "description": "List users" } }
      }
    },
    "/api/users/{username}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get public profile by username",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/api/users/{username}/posts": {
      "get": {
        "summary": "List posts by username (public)",
        "tags": ["Users"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "in": "query",
            "name": "page",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": { "type": "integer", "default": 20, "maximum": 50 }
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "User not found" }
        }
      }
    },
    "/api/users/{username}/followers": {
      "get": {
        "summary": "Get followers",
        "tags": ["Follow"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "User not found" }
        }
      }
    },
    "/api/users/{username}/following": {
      "get": {
        "summary": "Get following",
        "tags": ["Follow"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "User not found" }
        }
      }
    },
    "/api/posts": {
      "post": {
        "summary": "Create a post (upload image + caption)",
        "tags": ["Posts"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["image"],
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "JPG/PNG/WEBP (max 5MB)"
                  },
                  "caption": { "type": "string", "example": "Hello world!" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created" },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/posts/{id}": {
      "get": {
        "summary": "Get post detail",
        "tags": ["Posts"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "Not found" }
        }
      },
      "delete": {
        "summary": "Delete my own post",
        "tags": ["Posts"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Deleted" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/api/likes/{postId}": {
      "post": {
        "summary": "Like a post (alias)",
        "tags": ["Likes"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "Post not found" }
        }
      },
      "delete": {
        "summary": "Unlike a post (alias)",
        "tags": ["Likes"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "Post not found" }
        }
      }
    },
    "/api/posts/{id}/comments": {
      "post": {
        "summary": "Like a post",
        "tags": ["Likes"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "Post not found" }
        }
      },
      "delete": {
        "summary": "Unlike a post",
        "tags": ["Likes"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "Post not found" }
        }
      }
    },
    "/api/posts/{id}/comments": {
      "get": {
        "summary": "Get comments",
        "tags": ["Comments"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "Post not found" }
        }
      },
      "post": {
        "summary": "Add comment",
        "tags": ["Comments"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["comment"],
                "properties": {
                  "comment": { "type": "string", "example": "Great post!" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created" },
          "404": { "description": "Post not found" }
        }
      }
    },
    "/api/follows/{username}": {
      "post": {
        "summary": "Follow a user",
        "tags": ["Follow"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "User not found" }
        }
      },
      "delete": {
        "summary": "Unfollow a user",
        "tags": ["Follow"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "User not found" }
        }
      }
    },
    "/api/saves/{postId}": {
      "post": {
        "summary": "Save a post (bookmark)",
        "tags": ["Saves"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Saved" },
          "404": { "description": "Post not found" }
        }
      },
      "delete": {
        "summary": "Unsave a post",
        "tags": ["Saves"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Unsaved" },
          "404": { "description": "Post not found" }
        }
      }
    },
    "/api/users/{username}/saves": {
      "get": {
        "summary": "Get user's saved posts",
        "tags": ["Saves"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "in": "query",
            "name": "page",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": { "type": "integer", "default": 20, "maximum": 50 }
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "User not found" }
        }
      }
    },
    "/api/comments/{id}": {
      "delete": {
        "summary": "Delete my comment",
        "tags": ["Comments"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Deleted" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Comment not found" }
        }
      }
    },
    "/api/me/posts": {
      "get": {
        "summary": "Get posts created by the authenticated user",
        "tags": ["Posts"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": { "type": "integer", "default": 20, "maximum": 50 }
          }
        ],
        "responses": {
          "200": { "description": "OK" }
        }
      }
    },
    "/api/users/{username}/likes": {
      "get": {
        "summary": "List posts that this user has liked (public)",
        "tags": ["Likes"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "in": "query",
            "name": "page",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": { "type": "integer", "default": 20, "maximum": 50 }
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "User not found" }
        }
      }
    },
    "/api/posts/{id}/likes": {
      "get": {
        "summary": "List users who liked a post",
        "tags": ["Likes"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "Post not found" }
        }
      }
    },
    "/api/me/likes": {
      "get": {
        "summary": "Get posts I liked",
        "tags": ["Likes"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": { "type": "integer", "default": 20, "maximum": 50 }
          }
        ],
        "responses": {
          "200": { "description": "OK" }
        }
      }
    },
    "/api/me/followers": {
      "get": {
        "summary": "List my followers",
        "tags": ["Follow"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "OK" }
        }
      }
    },
    "/api/me/following": {
      "get": {
        "summary": "List users I follow",
        "tags": ["Follow"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "OK" }
        }
      }
    },
    "/api/posts/explore": {
      "get": {
        "summary": "Explore all posts (like Instagram explore)",
        "tags": ["Posts"],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": { "type": "integer", "default": 20, "maximum": 50 }
          }
        ],
        "responses": {
          "200": { "description": "OK" }
        }
      }
    }
  }
}